add error directory in page directory move error pages into /error delete data directory (only used in frontend)
