--- src/frontend/webappService/src/data/data.json
@@ -1,37 +1,0 @@
-{
-    "users": [
-        {
-            "id": 0,
-            "firstname": "Aïna",
-            "lastname": "Dirou",
-            "age": 22,
-            "job": "Frontend Developper",
-            "office": "Brest",
-            "department": "Software Engineering",
-            "username": "aina",
-            "password": "aina"
-        },
-        {
-            "id": 1,
-            "firstname": "Chloé",
-            "lastname": "Benoît",
-            "age": 23,
-            "job": "Network engineer",
-            "office": "Brest",
-            "department": "Network Engineering",
-            "username": "chloe",
-            "password": "chloe"
-        },
-        {
-            "id": 2,
-            "firstname": "Antoine",
-            "lastname": "Dagorn",
-            "age": 23,
-            "job": "Project manager",
-            "office": "New York",
-            "department": "Network Engineering",
-            "username": "antoine",
-            "password": "antoine"
-        }
-    ]
-}
