+++ src/frontend/webappService/src/pages/errors/Error403.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
+import Navbar from '../../components/Navbar'
 import { Link } from 'react-router-dom'
 
 function Error403({ pageLink = "/", message = "Go back to login" }) {
