+++ src/frontend/webappService/src/pages/errors/Error404.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
+import Navbar from '../../components/Navbar'
 import { Link } from 'react-router-dom'
 
 function Error404() {
